refactor: GpuServerRegisterForm business logic을 hook으로 분리
